feat: refactor SSL structure
