+++ src/pages/ssl/components/CertificateForm/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Form, Input } from 'antd';
 import { useIntl } from 'umi';
 import { FormInstance } from 'antd/lib/form';
-import { FormData } from '../..';
+import { FormData } from '../../Create';
 
 interface CertificateFormProps {
   mode: 'EDIT' | 'VIEW';
