+++ src/pages/ssl/components/Step3/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Form, Button } from 'antd';
 import { create as createSSL } from '@/pages/ssl/service';
 import CertificateForm from '../CertificateForm';
-import { StepProps } from '../..';
+import { StepProps } from '../../Create';
 
 const Step3: React.FC<StepProps> = ({ data, onStepChange }) => {
   const [form] = Form.useForm();
